@@ -2,6 +2,6 @@ import { ApplicationContract } from '../application/ApplicationContract';
 
 export interface ProviderContract {
     app: ApplicationContract;
-    boot: () => void;
+    boot: () => Promise<any>;
     register?: () => void;
 }
