@@ -1,3 +1,3 @@
-import Application from '../body/app/Application';
+import { Application } from '../body/app/Application';
 
-const application = Application.getInstance();
+const application = new Application();
